+++ 10주차/BookShop/controller/orderController.js
@@ -1,4 +1,5 @@
-// const conn = require('../mariadb');
+const ensureAuthorization = require('../auth.js');
+var jwt = require('jsonwebtoken');
 const mariadb = require('mysql2/promise');
 const { StatusCodes } = require('http-status-codes');
 
@@ -11,37 +12,49 @@ const order = async (req, res) => {
         dateStrings : true
     });
 
-    const {items, delivery, totalQuantity, totalPrice, userId, firstBookTitle} = req.body;
+    let authorization = ensureAuthorization(req, res);
 
-    // delivery 테이블 삽입
-    let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?)`;
-    let values = [delivery.address, delivery.receiver, delivery.contact];
-    let [results] = await conn.execute(sql, values);
-    let delivery_id = results.insertId;
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        const {items, delivery, totalQuantity, totalPrice, firstBookTitle} = req.body;
 
-    // orders 테이블 삽입
-    sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
-            VALUES (?, ?, ?, ?, ?)`;
-    values = [firstBookTitle, totalQuantity, totalPrice, userId, delivery_id];
-    [results] = await conn.execute(sql, values);
-    let order_id = results.insertId;
-
-    // items를 가지고 장바구니에서 book_id, quantity 조회 
-    sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
-    let [orderItems, fields] = await conn.query(sql, [items]); // select할 땐 [rows, fields] 로 받아줘야 함
-
-    // orderedBook 데이터 삽입
-    sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
-    values = []; // 초기화
-    orderItems.forEach((item) => {
-        values.push ([order_id, item.book_id, item.quantity]); // 배열 안에 배열이 들어감 (2차원 배열)
-    });
-    results = await conn.query(sql, [values]);
-        
-    // 장바구니에서 주문한 상품 삭제
-    let result = await deleteCartItems(conn, items);
+        // delivery 테이블 삽입
+        let sql = `INSERT INTO delivery (address, receiver, contact) VALUES (?, ?, ?)`;
+        let values = [delivery.address, delivery.receiver, delivery.contact];
+        let [results] = await conn.execute(sql, values);
+        let delivery_id = results.insertId;
     
-    return res.status(StatusCodes.OK).json(result);
+        // orders 테이블 삽입
+        sql = `INSERT INTO orders (book_title, total_quantity, total_price, user_id, delivery_id) 
+                VALUES (?, ?, ?, ?, ?)`;
+        values = [firstBookTitle, totalQuantity, totalPrice, authorization.id, delivery_id];
+        [results] = await conn.execute(sql, values);
+        let order_id = results.insertId;
+    
+        // items를 가지고 장바구니에서 book_id, quantity 조회 
+        sql = `SELECT book_id, quantity FROM cartItems WHERE id IN (?)`;
+        let [orderItems, fields] = await conn.query(sql, [items]); // select할 땐 [rows, fields] 로 받아줘야 함
+    
+        // orderedBook 데이터 삽입
+        sql = `INSERT INTO orderedBook (order_id, book_id, quantity) VALUES ?`;
+        values = []; // 초기화
+        orderItems.forEach((item) => {
+            values.push ([order_id, item.book_id, item.quantity]); // 배열 안에 배열이 들어감 (2차원 배열)
+        });
+        results = await conn.query(sql, [values]);
+            
+        // 장바구니에서 주문한 상품 삭제
+        let result = await deleteCartItems(conn, items);
+        
+        return res.status(StatusCodes.OK).json(result);
+    }
 };
 
 const deleteCartItems = async (conn, items) => {
@@ -52,42 +65,66 @@ const deleteCartItems = async (conn, items) => {
 }; 
 
 const getOrders = async (req, res) => {
-    const conn = await mariadb.createConnection({
-        host : 'localhost',
-        user : 'root',
-        password : 'root',
-        database : 'Bookshop',
-        dateStrings : true
-    });
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = `SELECT orders.id, created_at, address, receiver, contact, book_title, total_quantity, total_price
-                FROM orders LEFT JOIN delivery
-                ON orders.delivery_id = delivery.id`;
-
-    let [rows, fields] = await conn.query(sql);
-
-    return res.status(StatusCodes.OK).json(rows);
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        const conn = await mariadb.createConnection({
+            host : 'localhost',
+            user : 'root',
+            password : 'root',
+            database : 'Bookshop',
+            dateStrings : true
+        });
+    
+        let sql = `SELECT orders.id, created_at, address, receiver, contact, book_title, total_quantity, total_price
+                    FROM orders LEFT JOIN delivery
+                    ON orders.delivery_id = delivery.id`;
+    
+        let [rows, fields] = await conn.query(sql);
+    
+        return res.status(StatusCodes.OK).json(rows);
+    }
 };
 
 const getOrderDetail = async (req, res) => {
-    const {id} = req.params;
+    let authorization = ensureAuthorization(req, res);
 
-    const conn = await mariadb.createConnection({
-        host : 'localhost',
-        user : 'root',
-        password : 'root',
-        database : 'Bookshop',
-        dateStrings : true
-    });
-
-    let sql = `SELECT book_id, title, author, price, quantity
-                FROM orderedBook LEFT JOIN books
-                ON orderedBook.book_id = books.id
-                WHERE order_id = ?`;
-
-    let [rows, fields] = await conn.query(sql, [id]);
-
-    return res.status(StatusCodes.OK).json(rows);
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        const orderId = req.params.id;
+    
+        const conn = await mariadb.createConnection({
+            host : 'localhost',
+            user : 'root',
+            password : 'root',
+            database : 'Bookshop',
+            dateStrings : true
+        });
+    
+        let sql = `SELECT book_id, title, author, price, quantity
+                    FROM orderedBook LEFT JOIN books
+                    ON orderedBook.book_id = books.id
+                    WHERE order_id = ?`;
+    
+        let [rows, fields] = await conn.query(sql, [orderId]);
+    
+        return res.status(StatusCodes.OK).json(rows);
+    }
 };
 
 module.exports =  {
